move services/stateful-service.ts to services/core/
