+++ app/services/customization/customization.ts
@@ -1,6 +1,6 @@
 import { Subject } from 'rxjs';
 import { PersistentStatefulService } from '../persistent-stateful-service';
-import { mutation } from '../stateful-service';
+import { mutation } from '../core/stateful-service';
 import {
   ICustomizationServiceApi,
   ICustomizationServiceState,
