+++ app/services/hotkeys.ts
@@ -4,7 +4,7 @@ import { SourcesService } from 'services/sources';
 import { TransitionsService } from 'services/transitions';
 import { KeyListenerService } from 'services/key-listener';
 import { Inject } from 'services/core/injector';
-import { StatefulService, mutation, ServiceHelper } from 'services/stateful-service';
+import { StatefulService, mutation, ServiceHelper } from 'services/core/stateful-service';
 import { defer } from 'lodash';
 import { $t } from 'services/i18n';
