+++ app/services/i18n/i18n.test.ts
@@ -10,7 +10,7 @@ jest.mock('electron', () => ({
     }
   }
 }));
-jest.mock('services/stateful-service');
+jest.mock('services/core/stateful-service');
 jest.mock('services/core/injector');
 jest.mock('../../../obs-api', () => ({
   Global: {},
