+++ app/services/informations/informations.ts
@@ -1,5 +1,5 @@
 import { Inject } from 'services/core/injector';
-import { StatefulService, mutation } from 'services/stateful-service';
+import { StatefulService, mutation } from 'services/core/stateful-service';
 import { HostsService } from 'services/hosts';
 import { parseString } from 'xml2js';
 import { handleErrors } from 'util/requests';
