+++ app/services/nicolive-program/nicolive-comment-filter.ts
@@ -1,4 +1,4 @@
-import { StatefulService, mutation } from 'services/stateful-service';
+import { StatefulService, mutation } from 'services/core/stateful-service';
 import { NicoliveClient, isOk } from './NicoliveClient';
 import { FilterRecord } from './ResponseTypes';
 import { Inject } from 'services/core/injector';
