+++ app/services/nicolive-program/nicolive-comment-viewer.ts
@@ -1,6 +1,6 @@
 import { Inject } from 'services/core/injector';
 import { NicoliveProgramService } from 'services/nicolive-program/nicolive-program';
-import { StatefulService, mutation } from 'services/stateful-service';
+import { StatefulService, mutation } from 'services/core/stateful-service';
 import { Subscription, EMPTY, Observable, of } from 'rxjs';
 import {
   map,
