+++ app/services/nicolive-program/nicolive-program-selector.ts
@@ -1,4 +1,4 @@
-import { StatefulService, mutation } from 'services/stateful-service';
+import { StatefulService, mutation } from 'services/core/stateful-service';
 import { NicoliveClient, isOk } from './NicoliveClient';
 import { OnairChannelData } from './ResponseTypes';
 import { NicoliveFailure, openErrorDialogFromFailure } from './NicoliveFailure';
