+++ app/services/notifications/notifications.ts
@@ -1,5 +1,5 @@
 import { Inject } from '../core/injector';
-import { mutation } from '../stateful-service';
+import { mutation } from '../core/stateful-service';
 import { PersistentStatefulService } from 'services/persistent-stateful-service';
 import { Subject } from 'rxjs';
 import { WindowsService } from 'services/windows';
