+++ app/services/onboarding.ts
@@ -1,4 +1,4 @@
-import { StatefulService, mutation } from './stateful-service';
+import { StatefulService, mutation } from './core/stateful-service';
 import { NavigationService } from './navigation';
 import { UserService } from './user';
 import { Inject } from './core/injector';
