+++ app/services/patch-notes/index.ts
@@ -1,5 +1,5 @@
 import { PersistentStatefulService } from 'services/persistent-stateful-service';
-import { mutation } from 'services/stateful-service';
+import { mutation } from 'services/core/stateful-service';
 import electron from 'electron';
 import Util from 'services/utils';
 import { notes } from './notes';
