+++ app/services/performance/performance.test.ts
@@ -7,7 +7,7 @@ const setup = createSetupFunction({
   },
 });
 
-jest.mock('services/stateful-service');
+jest.mock('services/core/stateful-service');
 jest.mock('services/core/injector');
 jest.mock('services/settings', () => ({}));
 jest.mock('services/customization', () => ({}));
