+++ app/services/platforms/niconico.test.ts
@@ -14,7 +14,7 @@ const setup = createSetupFunction({
   },
 });
 
-jest.mock('services/stateful-service');
+jest.mock('services/core/stateful-service');
 jest.mock('services/core/injector');
 jest.mock('services/streaming', () => ({}));
 jest.mock('services/user', () => ({}));
