+++ app/services/scene-collections/state.test.ts
@@ -1,5 +1,5 @@
 import { createSetupFunction } from 'util/test-setup';
-jest.mock('services/stateful-service');
+jest.mock('services/core/stateful-service');
 jest.mock('services/core/injector');
 
 const setup = createSetupFunction();
