+++ app/services/sources/monitor-capture-cropping.ts
@@ -2,7 +2,7 @@ import electron, { BrowserWindow } from 'electron';
 import { Inject } from 'services/core/injector';
 import { WindowsService } from 'services/windows';
 import { ISourcesServiceApi } from 'services/sources';
-import { StatefulService, mutation } from '../stateful-service';
+import { StatefulService, mutation } from '../core/stateful-service';
 import { ScalableRectangle, ResizeBoxPoint } from 'util/ScalableRectangle';
 import { SceneItem } from '../scenes';
