+++ app/services/streaming/streaming.test.ts
@@ -5,7 +5,7 @@ import { createSetupFunction } from 'util/test-setup';
 
 function noop(..._args: any[]) { }
 
-jest.mock('services/stateful-service');
+jest.mock('services/core/stateful-service');
 jest.mock('services/core/injector');
 jest.mock('../../../obs-api', () => ({
   NodeObs: {
