+++ app/services/troubleshooter/troubleshooter.ts
@@ -1,4 +1,4 @@
-import { mutation } from '../stateful-service';
+import { mutation } from '../core/stateful-service';
 import { PersistentStatefulService } from 'services/persistent-stateful-service';
 import { IObsInput, IObsNumberInputValue, TObsFormData } from 'components/obs/inputs/ObsInput';
 import { ITroubleshooterServiceApi, ITroubleshooterSettings, TIssueCode } from './troubleshooter-api';
