+++ app/store/index.ts
@@ -2,7 +2,7 @@ import Vue from 'vue';
 import Vuex, { Store } from 'vuex';
 import _ from 'lodash';
 import electron from 'electron';
-import { getModule, StatefulService } from '../services/stateful-service';
+import { getModule, StatefulService } from '../services/core/stateful-service';
 import { ServicesManager } from '../services-manager';
 import { IMutation } from 'services/jsonrpc';
 import Util from 'services/utils';
